Add "min regi" page
Added regi form, log, and information about the users regi. This needs to be tweaked. The database is not yet updated to support these changes. `uploadRegiImages` is then a placeholder.
Removed the firebase email form `useAuth` since it should already be in the user data.
Updated Typescript version to fix issue with zod.
Also refactor `useAuth` and `AuthContext` to typescript and to reduce nesting.